alerting/requirements.txt
google-cloud-bigquery
sendgrid